Create SCOUSE_stage_3.pro
To speed up Stage 3, I made similar changes to those that I made in the create_coordfiles for Stage 1. Instead of having a where function in a for loop, I do the coordinate matching between different arrays outside of the for loop. To do this matching, I use this routine: https://www.harrisgeospatial.com/docs/matchall_2d.html  Thank you!